refactor: 닉네임 관련 API Migration 및 UserRole 관련 변경
- 닉네임 관련 API Migration
  : member -> profile

- UserRole 관련 변경 사항
  : UserRole (일반회원) 을 입력받지 않고, Default 로 일반회원으로 설정되게끔 변경
