--- src/main/java/com/unionclass/memberservice/client/profile/ProfileServiceClient.java
@@ -1,9 +1,0 @@
-package com.unionclass.memberservice.client.profile;
-
-import org.springframework.cloud.openfeign.FeignClient;
-
-@FeignClient(name = "profile-service")
-public interface ProfileServiceClient {
-
-    public void registerNickname();
-}
