+++ src/main/java/com/unionclass/memberservice/common/response/ResponseMessage.java
@@ -13,7 +13,6 @@ public enum ResponseMessage {
     SUCCESS_VERIFY_EMAIL_CODE("메일 인증코드 검증에 성공하였습니다."),
     SUCCESS_CHECK_EMAIL_DUPLICATE("이메일 중복 검사에 성공하였습니다."),
     SUCCESS_CHECK_LOGIN_ID_DUPLICATE("아이디 중복 검사에 성공하였습니다."),
-    SUCCESS_CHECK_NICKNAME_DUPLICATE("닉네임 중복 검사에 성공하였습니다."),
     SUCCESS_SEND_TEMPORARY_PASSWORD("임시 비밀번호 발급에 성공하였습니다."),
     SUCCESS_CHANGE_PASSWORD("비밀번호 변경에 성공하였습니다."),
     SUCCESS_SIGN_IN_WITH_OAUTH("소셜 로그인에 성공하였습니다."),
@@ -25,7 +24,6 @@ public enum ResponseMessage {
     SUCCESS_DELETE_AGREEMENT("약관동의 항목 삭제에 성공하였습니다."),
     SUCCESS_REGISTER_MEMBER_AGREEMENT("회원 약관동의 여부 등록에 성공하였습니다."),
     SUCCESS_UPDATE_MEMBER_AGREEMENT("회원 약관동의 여부 변경에 성공하였습니다."),
-    SUCCESS_CHANGE_NICKNAME("닉네임 변경에 성공하였습니다."),
     SUCCESS_GET_MY_INFO("내 정보 조회에 성공하였습니다.")
     ;
