+++ src/main/java/com/unionclass/memberservice/domain/auth/presentation/AuthController.java
@@ -1,24 +1,25 @@
 package com.unionclass.memberservice.domain.auth.presentation;
 
+import com.unionclass.memberservice.common.response.BaseResponseEntity;
+import com.unionclass.memberservice.common.response.ResponseMessage;
 import com.unionclass.memberservice.domain.auth.application.AuthService;
 import com.unionclass.memberservice.domain.auth.dto.in.GetLoginIdReqDto;
-import com.unionclass.memberservice.domain.auth.dto.in.GetNicknameReqDto;
 import com.unionclass.memberservice.domain.auth.dto.in.SignInReqDto;
 import com.unionclass.memberservice.domain.auth.dto.in.SignUpReqDto;
 import com.unionclass.memberservice.domain.auth.vo.in.GetLoginIdReqVo;
-import com.unionclass.memberservice.domain.auth.vo.in.GetNicknameReqVo;
 import com.unionclass.memberservice.domain.auth.vo.in.SignInReqVo;
 import com.unionclass.memberservice.domain.auth.vo.in.SignUpReqVo;
 import com.unionclass.memberservice.domain.auth.vo.out.SignInResVo;
-import com.unionclass.memberservice.common.response.BaseResponseEntity;
-import com.unionclass.memberservice.common.response.ResponseMessage;
 import com.unionclass.memberservice.domain.email.dto.in.EmailReqDto;
 import com.unionclass.memberservice.domain.email.vo.in.EmailReqVo;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
@@ -35,7 +36,6 @@ public class AuthController {
      * 2. 로그인
      * 3. 이메일 중복 검사
      * 4. 아이디 중복 검사
-     * 5. 닉네임 중복 검사
      */
 
     /**
@@ -150,35 +150,5 @@ public class AuthController {
     ) {
         authService.checkLoginIdDuplicate(GetLoginIdReqDto.from(getLoginIdReqVo));
         return new BaseResponseEntity<>(ResponseMessage.SUCCESS_CHECK_LOGIN_ID_DUPLICATE.getMessage());
-    }
-
-    /**
-     * 5. 닉네임 중복 검사
-     *
-     * @param nicknameReqVo
-     * @return
-     */
-    @Operation(
-            summary = "닉네임 중복 검사",
-            description = """
-                    사용자가 입력한 닉네임(nickname)이 이미 사용 중인지 확인합니다.
-
-                    [요청 조건]
-                    - nickname: 필수 입력, 공백 불가
-                
-                    [처리 방식]
-                    - 닉네임으로 회원을 조회하여 존재 여부를 확인합니다.
-                    - 이미 존재하는 경우 예외를 발생시키고, 존재하지 않으면 성공 응답을 반환합니다.
-                
-                    [예외 코드]
-                    - NICKNAME_ALREADY_EXISTS: 중복된 닉네임일 경우
-                    """
-    )
-    @PostMapping("/nickname/check-duplicate")
-    public BaseResponseEntity<Void> checkNicknameDuplicate(
-            @Valid @RequestBody GetNicknameReqVo nicknameReqVo
-    ) {
-        authService.checkNicknameDuplicate(GetNicknameReqDto.from(nicknameReqVo));
-        return new BaseResponseEntity<>(ResponseMessage.SUCCESS_CHECK_NICKNAME_DUPLICATE.getMessage());
     }
 }
