--- src/main/java/com/unionclass/memberservice/domain/auth/vo/in/GetNicknameReqVo.java
@@ -1,11 +1,0 @@
-package com.unionclass.memberservice.domain.auth.vo.in;
-
-import jakarta.validation.constraints.NotBlank;
-import lombok.Getter;
-
-@Getter
-public class GetNicknameReqVo {
-
-    @NotBlank(message = "닉네임을 입력해주세요.")
-    private String nickname;
-}
