+++ src/main/java/com/unionclass/memberservice/domain/member/application/MemberService.java
@@ -1,6 +1,5 @@
 package com.unionclass.memberservice.domain.member.application;
 
-import com.unionclass.memberservice.domain.member.dto.in.ChangeNicknameReqDto;
 import com.unionclass.memberservice.domain.member.dto.in.ChangePasswordReqDto;
 import com.unionclass.memberservice.domain.member.dto.in.ResetPasswordReqDto;
 import com.unionclass.memberservice.domain.member.dto.out.GetMyInfoResDto;
@@ -9,6 +8,5 @@ public interface MemberService {
 
     void changePassword(ChangePasswordReqDto changePasswordReqDto);
     void resetPasswordWithTemporary(ResetPasswordReqDto resetPasswordReqDto);
-    void changeNickname(ChangeNicknameReqDto changeNicknameReqDto);
     GetMyInfoResDto getMyInfo(String memberUuid);
 }
