+++ src/main/java/com/unionclass/memberservice/domain/member/infrastructure/MemberRepository.java
@@ -10,5 +10,4 @@ public interface MemberRepository extends JpaRepository<Member, Long> {
     Optional<Member> findByLoginId(String loginId);
     Optional<Member> findByMemberUuid(String memberUuid);
     Optional<Member> findByEmail(String email);
-    Optional<Member> findByNickname(String nickname);
 }
