--- src/main/java/com/unionclass/memberservice/domain/member/vo/in/ChangeNicknameReqVo.java
@@ -1,9 +1,0 @@
-package com.unionclass.memberservice.domain.member.vo.in;
-
-import lombok.Getter;
-
-@Getter
-public class ChangeNicknameReqVo {
-
-    private String nickname;
-}
